pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "secret-detection-framework"
version = "1.0.0"
description = "A production-grade platform for detecting and rotating secrets in Git repositories"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Security Team", email = "security@secret-framework.io"}
]
maintainers = [
    {name = "Security Team", email = "security@secret-framework.io"}
]
keywords = [
    "security",
    "secrets",
    "credentials",
    "git",
    "aws",
    "azure",
    "github",
    "devsecops",
    "secret-detection",
    "credential-rotation"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Topic :: Software Development :: Version Control :: Git",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Typing :: Typed"
]

dependencies = [
    "gitpython>=3.1.31",
    "pygithub>=2.1.1",
    "boto3>=1.28.0",
    "azure-identity>=1.14.0",
    "azure-mgmt-authorization>=4.0.0",
    "azure-graphrbac>=0.61.1",
    "regex>=2023.6.3",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.3.0",
    "pydantic-settings>=2.0.3",
    "click>=8.1.7",
    "rich>=13.5.2",
    "tabulate>=0.9.0",
    "streamlit>=1.26.0",
    "plotly>=5.16.1",
    "pandas>=2.1.0",
    "cryptography>=41.0.3",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.1",
    "aiofiles>=23.2.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-xdist>=3.3.1",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "flake8-docstrings>=1.7.0",
    "mypy>=1.5.1",
    "isort>=5.12.0",
    "bandit>=1.7.5",
    "safety>=2.3.5",
    "pre-commit>=3.3.3"
]
docs = [
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0"
]
test = [
    "pytest>=7.4.2",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "moto[all]>=4.2.2",
    "responses>=0.23.3"
]

[project.urls]
Homepage = "https://github.com/your-org/secret-framework"
Documentation = "https://secret-framework.readthedocs.io"
Repository = "https://github.com/your-org/secret-framework"
Issues = "https://github.com/your-org/secret-framework/issues"
Changelog = "https://github.com/your-org/secret-framework/blob/main/CHANGELOG.md"

[project.scripts]
secretctl = "cli.secretctl:cli"

[tool.setuptools]
packages = ["detectors", "scanners", "rotators", "api", "cli", "dashboard", "reporting"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.j2"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
split_on_trailing_comma = true
skip_glob = ["*/migrations/*", "*/venv/*", "*/env/*"]
known_third_party = ["fastapi", "pydantic", "boto3", "azure", "github", "streamlit"]
known_first_party = ["detectors", "scanners", "rotators", "api", "cli", "dashboard"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=.",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-n auto"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "aws: Tests requiring AWS credentials",
    "azure: Tests requiring Azure credentials",
    "github: Tests requiring GitHub token",
    "docker: Tests requiring Docker"
]
log_cli = false
log_cli_level = "INFO"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "*/build/*",
    "*/dist/*",
    "setup.py"
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
    "@abc.abstractmethod",
    "if TYPE_CHECKING:",
    "\\.\\.\\."
]

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_context = true
show_column_numbers = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "boto3.*",
    "botocore.*",
    "azure.*",
    "github.*",
    "git.*",
    "streamlit.*",
    "plotly.*",
    "pandas.*"
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "venv", "env", ".venv"]
skips = ["B101", "B601"]

[tool.pylint.main]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]
exclude = [
    ".git",
    "__pycache__",
    "dist",
    "build",
    "*.egg-info"
]

